oci support: refactor multi upload for file
